requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
scikit-learn
TPOT
dl8.5
gplearn
symreg
auto-sklearn